+++ src/test/java/nextstep/refactoring/ladder/concrete/LadderLineTest.java
@@ -1,5 +1,8 @@
-package nextstep.refactoring.ladder;
+package nextstep.refactoring.ladder.concrete;
 
+import nextstep.refactoring.ladder.concrete.Direction;
+import nextstep.refactoring.ladder.concrete.LadderLine;
+import nextstep.refactoring.ladder.concrete.LadderPoint;
 import nextstep.refactoring.ladder.engine.Line;
 import nextstep.refactoring.ladder.engine.Position;
 import org.junit.jupiter.api.DisplayName;
