+++ src/test/java/nextstep/refactoring/ladder/concrete/LadderPointTest.java
@@ -1,5 +1,7 @@
-package nextstep.refactoring.ladder;
+package nextstep.refactoring.ladder.concrete;
 
+import nextstep.refactoring.ladder.concrete.Direction;
+import nextstep.refactoring.ladder.concrete.LadderPoint;
 import nextstep.refactoring.ladder.engine.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
