+++ src/test/java/nextstep/refactoring/ladder/concrete/TestLadder.java
@@ -1,5 +1,9 @@
-package nextstep.refactoring.ladder;
+package nextstep.refactoring.ladder.concrete;
 
+import nextstep.refactoring.ladder.concrete.Direction;
+import nextstep.refactoring.ladder.concrete.LadderImpl;
+import nextstep.refactoring.ladder.concrete.LadderLine;
+import nextstep.refactoring.ladder.concrete.LadderPoint;
 import nextstep.refactoring.ladder.engine.Ladder;
 import nextstep.refactoring.ladder.engine.Line;
